Refactor codebase
refactor code base for better modularity and maintainability
