--- ToDo
@@ -1,4 +1,0 @@
-Make is workable without zapier webhook
-Add readme for deloymnet
-Create a telegram group for the community
-Update license
